設定セクションの説明文用コンポーネントを分離した
